Version 1/requirements.txt
moviepy==1.0.3
arabic-reshaper==3.0.0
python-bidi==0.4.2
ffmpeg==1.4
ffmpeg-downloader==0.2.0
ffmpeg-python==0.2.0
imageio-ffmpeg==0.4.8
pysrt==1.1.2

